+++ cli/cli-core/src/test/java/io/telicent/smart/cache/cli/commands/options/TestKafkaOptions.java
@@ -50,7 +50,7 @@ public class TestKafkaOptions extends AbstractCommandTests {
 
     @AfterMethod
     @Override
-    public void testCleanup() throws InterruptedException {
+    public void testCleanup() {
         super.testCleanup();
         Configurator.reset();
     }