+++ cli/cli-core/src/test/java/io/telicent/smart/cache/cli/options/TestOffsetStoreOptions.java
@@ -18,15 +18,46 @@ package io.telicent.smart.cache.cli.options;
 import io.telicent.smart.cache.cli.commands.AbstractCommandTests;
 import io.telicent.smart.cache.cli.commands.SmartCacheCommand;
 import io.telicent.smart.cache.cli.commands.SmartCacheCommandTester;
+import io.telicent.smart.cache.configuration.Configurator;
+import io.telicent.smart.cache.configuration.sources.PropertiesSource;
 import io.telicent.smart.cache.sources.offsets.OffsetStore;
 import io.telicent.smart.cache.sources.offsets.file.JsonOffsetStore;
 import io.telicent.smart.cache.sources.offsets.file.YamlOffsetStore;
 import org.apache.jena.atlas.lib.FileOps;
 import org.testng.Assert;
 import org.testng.annotations.AfterClass;
+import org.testng.annotations.AfterMethod;
+import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
-public class TestOffsetStoreOptions  extends AbstractCommandTests {
+import java.util.Properties;
+
+public class TestOffsetStoreOptions extends AbstractCommandTests {
+
+    @Override
+    @BeforeClass
+    public void setup() {
+        super.setup();
+        Configurator.reset();
+    }
+
+    @Override
+    @AfterMethod
+    public void testCleanup() {
+        super.testCleanup();
+        Configurator.reset();
+    }
+
+    @Override
+    @AfterClass
+    public void teardown() {
+        super.teardown();
+        Configurator.reset();
+        FileOps.deleteSilent("example.json");
+        FileOps.deleteSilent("example.yml");
+        FileOps.deleteSilent("example.yaml");
+    }
+
     @Test
     public void givenNoArguments_defaultToNull() {
         // Given
@@ -111,14 +142,23 @@ public class TestOffsetStoreOptions  extends AbstractCommandTests {
     }
 
     private static OffsetStore getOffsetStore() {
+        Assert.assertNotNull(SmartCacheCommandTester.getLastParseResult());
         return ((OffsetStoreOptionsCommand) SmartCacheCommandTester.getLastParseResult()
-                .getCommand()).offsetOptions.getOffsetStore();
+                                                                   .getCommand()).offsetOptions.getOffsetStore();
     }
 
-    @AfterClass
-    public static void afterClass() {
-        FileOps.deleteSilent("example.json");
-        FileOps.deleteSilent("example.yml");
-        FileOps.deleteSilent("example.yaml");
+    @Test
+    public void givenExternalConfig_whenCreatingOffsetStoreOptions_thenExternalConfigUsedAsDefaults() {
+        // Given
+        Properties properties = new Properties();
+        properties.put(CliEnvironmentVariables.OFFSETS_FILE, "target/offsets.yaml");
+        Configurator.setSingleSource(new PropertiesSource(properties));
+
+        // When
+        OffsetStoreOptions options = new OffsetStoreOptions();
+        OffsetStore store = options.getOffsetStore();
+
+        // Then
+        Assert.assertTrue(store instanceof YamlOffsetStore);
     }
 }