+++ cli/cli-debug/src/test/java/io/telicent/smart/cache/cli/commands/debug/DockerTestDebugCliSecureKafka.java
@@ -70,7 +70,7 @@ public class DockerTestDebugCliSecureKafka extends AbstractCommandTests {
 
     @AfterMethod
     @Override
-    public void testCleanup() throws InterruptedException {
+    public void testCleanup() {
         super.testCleanup();
         this.kafka.resetTestTopic();
         this.kafka.resetTopic(LiveReporter.DEFAULT_LIVE_TOPIC);